core/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
bandit==1.7.5
billiard==4.1.0
celery==5.3.4
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-elasticsearch-dsl==7.3
django-guardian==2.4.0
django-lifecycle==1.0.0
django-redis==5.4.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elastic-transport==8.4.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.1
flower==2.0.1
gitdb==4.0.10
GitPython==3.1.32
gprof2dot==2022.7.29
humanize==4.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.12.0
kombu==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.1
pbr==5.11.1
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
ruff==0.0.286
sentry-sdk==1.29.2
service-identity==23.1.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
stevedore==5.1.0
tornado==6.3.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
urlman==2.0.1
vine==5.0.0
wcwidth==0.2.8
zope.interface==6.0
